requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
sqlparse==0.5.5
django-filter==25.2
